requirements.txt
# Web & Scraping
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Async
asyncio-contextmanager==1.0.0

# Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
responses==0.14.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Optional: Browser automation (use if Phase 3 needs JS rendering)
playwright==1.40.0
